Move the cucumber code out into it's own gem
