+++ examples/cucumber-capybara/Gemfile
@@ -4,3 +4,4 @@ gem 'cucumber'
 gem 'rspec'
 gem 'capybara'
 gem 'sauce', :path => '../../'
+gem 'sauce-cucumber', :path => '../../gems/sauce-cucumber'
