+++ examples/cucumber-capybara/features/support/env.rb
@@ -1,6 +1,5 @@
 
 require 'capybara/cucumber'
-require 'sauce/capybara'
-require 'sauce/capybara/cucumber'
+require 'sauce/cucumber'
 
 Capybara.current_driver = :sauce
