+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 $:.unshift File.expand_path(File.dirname(__FILE__) + '/../lib')
 $:.unshift File.expand_path(File.dirname(__FILE__) + '/../gems/sauce-jasmine/lib')
+$:.unshift File.expand_path(File.dirname(__FILE__) + '/../gems/sauce-cucumber/lib')
 
 require 'sauce'
